Revert "fix robots at goal not recognized on small phones due to float rounding: Calculate cell size to fit the grid - use integer division to avoid rounding errors"
This reverts commit fbf234eaf02bb175b239a560b05a2ae70ea103dd.

+++ app/src/main/java/roboyard/ui/components/GameGridView.java
@@ -572,9 +572,9 @@ public class GameGridView extends View {
         int width = getMeasuredWidth();
         int height = getMeasuredHeight();
         
-        // Calculate cell size to fit the grid - use integer division to avoid rounding errors
-        int cellWidth = width / gridWidth;
-        int cellHeight = height / gridHeight;
+        // Calculate cell size to fit the grid
+        float cellWidth = (float) width / gridWidth;
+        float cellHeight = (float) height / gridHeight;
         cellSize = Math.min(cellWidth, cellHeight);
         
         // Calculate path rendering values based on cellSize for responsive design
